% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxc_bmi.R
\name{mxc_bmi}
\alias{mxc_bmi}
\title{MxC BMI Function}
\usage{
mxc_bmi(home, filename)
}
\arguments{
\item{home}{Parent directory path.}

\item{filename}{String name of file with the runs eg. "Baseline"}
}
\value{
Will write multiple csv files into the Folders with the experiments.
}
\description{
This function will run the BMI analysis portions of the Mexico Analysis.
}
\details{
This function will read in the folder with the runs, and read in each run. Then
call mxc_bmiind which will find the numbers/percentages of people OWO for each age and gender
grouping we are interested in.
mxc_bmi()
}
\examples{
mxc_bmi(home, "Baseline")
mxc_bmi(home, "Intervention 1")
}
\seealso{
{\code{\link{mxc_master}}, \code{\link{mxc_bmiind}}}
}
\author{
Sindiso Nyathi
}
