% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxc_bmiind.R
\name{mxc_bmiind}
\alias{mxc_bmiind}
\title{MxC BMI Ind}
\usage{
mxc_bmiind(age_run)
}
\arguments{
\item{age_run}{This is a run .csv file for a given gender age group.}
}
\value{
Returns a list with the values calculated for furhte rprocessing.
}
\description{
This function will run the BMI analysis for a single run.
}
\details{
This function will take in a run for a given age/gender group and find the number
average BMI and BMI percentile as well as the proporiton OWO at the beggining and end of the
simulation.
mxc_bmiind()
}
\examples{
mxc_bmiind(runF10_14)
mxc_bmiind(runM6_9)
}
\seealso{
{\code{\link{mxc_master}}, \code{\link{mxc_bmi}}}
}
\author{
Sindiso Nyathi
}
