% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxc_bmiindsd.R
\name{mxc_bmiindsd}
\alias{mxc_bmiindsd}
\title{MxC BMI Ind SD}
\usage{
mxc_bmiindsd(age_run)
}
\arguments{
\item{age_run}{This is a run .csv file for a given gender age group.}
}
\value{
Returns a list with the values calculated for furhte rprocessing.
}
\description{
This function will run the BMI Standard Deviation analysis for a single run.
}
\details{
This function will take in a run for a given age/gender group and find the SD of
average BMI and BMI percentile as well as the proporiton OWO at the beggining and end of the
simulation.
mxc_bmiind()
}
\examples{
mxc_bmiindsd(runF10_14)
mxc_bmiindsd(runM6_9)
}
\seealso{
{\code{\link{mxc_master}}, \code{\link{mxc_bmi}}, \code{\link{mxc_bmiind}}}
}
\author{
Sindiso Nyathi
}
