% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxc_dealsd.R
\name{mxc_dealsd}
\alias{mxc_dealsd}
\title{MxC Dealsd}
\usage{
mxc_dealsd(mr)
}
\arguments{
\item{mr}{File with the standard deviations.}
}
\value{
A file with the + or - value that will make the 95% confidence interval.
}
\description{
This function finds confidence intervals given standard deviations.
}
\details{
To be completed.
}
\examples{
mxc_dealsd(sdfile)
}
\seealso{
{\code{\link{mxc_master}}, \code{\link{mxc_mulruns}}, \code{\link{mxc_indrun}}}
}
\author{
Sindiso Nyathi
}
