% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxc_dealsdbmi.R
\name{mxc_dealsdbmi}
\alias{mxc_dealsdbmi}
\title{MxC Dealsd BMI}
\usage{
mxc_dealsdbmi(mr, ch)
}
\arguments{
\item{mr}{File with the standard deviations.}

\item{ch}{Columns we are interested in.}
}
\value{
A file with the + or - value that will make the 95% confidence interval.
}
\description{
This function finds confidence intervals given standard deviations for BMIs.
}
\details{
To be completed.
}
\examples{
mxc_dealsdbmi(sdfile, ch)
}
\seealso{
{\code{\link{mxc_master}}, \code{\link{mxc_bmi}}, \code{\link{mxc_bmiind}}}
}
\author{
Sindiso Nyathi
}
