% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxc_extrend.R
\name{mxc_extrend}
\alias{mxc_extrend}
\title{MxC Experiment Trend}
\usage{
mxc_extrend(home, filename)
}
\arguments{
\item{home}{Parent directory path.}

\item{filename}{String name of file with the runs eg. "Baseline"}
}
\value{
This function will save to the parent directorya file with the run averages for each Experiment.
}
\description{
This function will read in all the runs in an Experiment, average the bmi trend and save them to a
file in the parent directory for this run.
}
\details{
This function will read in a run, find the averages of the BMI trend values at each point
and save them for each run. Then at the end it will agerage the values for each of the N runs and save
the file in the parent directory.
}
\examples{
mxc_extrend(home, "Baseline")
mxc_extrend(home, "Intervention 1")
}
\seealso{
{\code{\link{mxc_master}}, \code{\link{mxc_mulruns}}, \code{\link{mxc_plot}}}
}
\author{
Sindiso Nyathi
}
