% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxc_indrun.R
\name{mxc_indrun}
\alias{mxc_indrun}
\title{MxC Indrun}
\usage{
mxc_indrun(home, results, run_number, filename, simulation_length)
}
\arguments{
\item{home}{Parent directory path.}

\item{results}{The run .csv file with the adolescents.}

\item{run_number}{The run number.}

\item{filename}{String name of file with the runs eg. "Baseline"}

\item{simulation_length}{Length of the simulation}
}
\value{
A file formatted, with required groupings and ages for the variables of interest.
}
\description{
This function will read in a single run and group adolescents by gender and age, and find variable averages
}
\details{
This function will aggregate a run into age and gender groupings, find the averages for
the Time and Events variables.This is done by taking the variable values for each gender/age grouping
and finding the means, for each of these. The Gender age groupings processed are m6-9yrs, f6 - 9yrs, m10-15yrs, f10-15yrs,
m15-18yrs, f15-18yrs, as well as male overall, and female overall and all overall.
}
\examples{
mxc_indrun(home, mxc_run2.csv, 2, "Baseline", 5)
mxc_indrun(home, mxc_run34.csv, 34, "Intervention 1", 5)
}
\seealso{
{\code{\link{mxc_master}}, \code{\link{mxc_mulruns}}, \code{\link{mxc_dealsd}}}
}
\author{
Sindiso Nyathi
}
