% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxc_indrunSD.R
\name{mxc_indrunSD}
\alias{mxc_indrunSD}
\title{MxC Indrun Standard Deviation}
\usage{
mxc_indrunSD(home, results, run_number, filename, simulation_length)
}
\arguments{
\item{home}{Parent directory path.}

\item{results}{The run .csv file with the adolescents.}

\item{run_number}{The run number.}

\item{filename}{String name of file with the runs eg. "Baseline"}

\item{simulation_length}{Length of the simulation}
}
\value{
A formatted file, similar to the file returned by mxc_mulruns, but with Standard Deviations inctead of means for the
required groupings and ages of the variables of interest.
}
\description{
This function is in many ways similar to mxc_indrun. It will read in a single run and group adolescents by gender and age,
but unlike mulruns, it will find the standard deviations instead of the means. The function only finds the SD of certain variables, not
all of them.
}
\details{
This function will aggregate a run into age and gender groupings, find the standard deviations for
the Time and Events variables.This is done by taking the variable values for each gender/age grouping
and finding the sd, for each of these. The Gender age groupings processed are m6-9yrs, f6 - 9yrs, m10-15yrs, f10-15yrs,
m15-18yrs, f15-18yrs, as well as male overall, and female overall and all overall.
}
\examples{
mxc_indrunSD(home, mxc_run2.csv, 2, "Baseline", 5)
mxc_indrunSD(home, mxc_run34.csv, 34, "Intervention 1", 5)
}
\seealso{
{\code{\link{mxc_master}}, \code{\link{mxc_mulruns}}, \code{\link{mxc_dealsd}}}
}
\author{
Sindiso Nyathi
}
