% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxc_plot.R
\name{mxc_plot}
\alias{mxc_plot}
\title{MxC Plot}
\usage{
mxc_plot(home, filename)
}
\arguments{
\item{home}{Parent directory path.}

\item{filename}{String name of file with the runs eg. "Baseline"}
}
\value{
This function will save to the disk several plots with the BMI trends for the age groups.
}
\description{
This  function will plot BMI trend.
}
\details{
This function will read in BMI 5 yr Trend and plot the results.
}
\examples{
mxc_plot(home, "Baseline")
mxc_plot(home, "Intervention 1")
}
\seealso{
{\code{\link{mxc_trend}}, \code{\link{mxc_mulruns}}, \code{\link{mxc_bmi}}}
}
\author{
Sindiso Nyathi
}
