% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxc_rename.R
\name{mxc_rename}
\alias{mxc_rename}
\title{MxC Rename}
\usage{
mxc_rename(home, foldername)
}
\arguments{
\item{home}{Parent directory with all the experiments.}

\item{foldername}{String name of folder with the runs eg. "Baseline"}
}
\value{
Nothing. Will rename the files in the folder.
}
\description{
This function will rename all the runs in the Experiment Folder specified.
}
\details{
Method: This function will take the names of all the runs in foldername and rename them from whatever their original names are
to mxc run_1.csv, mxc run_2.csv, mxc run_3.csv, . . . , mxc run_N.csv.
}
\examples{
mxc_rename(home, "Baseline")
mxc_master(home, "Intervention 1")
}
\seealso{
{\code{\link{mxc_mulruns}}, \code{\link{mxc_master}}, \code{\link{mxc_untar}}}
}
\author{
Sindiso Nyathi
}
