% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxc_trend.R
\name{mxc_trend}
\alias{mxc_trend}
\title{MxC Trend}
\usage{
mxc_trend(home, filename)
}
\arguments{
\item{home}{Parent directory path.}

\item{filename}{String name of file with the runs eg. "Baseline"}
}
\value{
This function will save to the disk several plots with the BMI trends for the age groups.
}
\description{
This  function will read in a all the runs, average the bmi trend for different groups and plot them.
}
\details{
This function will read in a run, find the averages of the BMI trend values at each point
and save them for each run for each group. Then at the end it will agerage the values for each of hte N runs
and plot that value in a graph.
}
\examples{
mxc_trend(home, "Baseline")
mxc_trend(home, "Intervention 1")
}
\seealso{
{\code{\link{mxc_master}}, \code{\link{mxc_mulruns}}, \code{\link{mxc_bmi}}}
}
\author{
Sindiso Nyathi
}
