% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxc_untar.R
\name{mxc_untar}
\alias{mxc_untar}
\title{MxC Untar}
\usage{
mxc_untar(home, foldername)
}
\arguments{
\item{home}{Parent directory with all the experiments.}

\item{foldername}{String name of file with the runs eg. "Baseline"}
}
\value{
For N runs originally  in .tar or .gz format, will return N files as .csv files.
}
\description{
This function will convert all the files in the run folder from .gz or .tar files to .csv files.
The function will not function if the files are compressed into any other format. It will only
work with .tar or .gz files.
}
\details{
Method: This function reads in each .tar or .gz file, converts it to a .csv file and saves the ouput
to the current folder.
}
\examples{
mxc_untar(home, "Baseline")
mxc_untar(home, "Experiment1")
}
\seealso{
{\code{\link{mxc_rename}}, \code{\link{mxc_master}}}
}
\author{
Sindiso Nyathi
}
